requirements_local.txt
# Pour la gestion multilingue (i18n)
Flask-Babel==3.1.0
# Requirements pour AcadCheck - Installation Locale
# Liste des packages nécessaires pour faire fonctionner AcadCheck localement

Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-WTF==1.1.1
Werkzeug==2.3.6
WTForms==3.0.1
python-dotenv==1.0.0
requests==2.31.0
PyPDF2==3.0.1
python-docx==0.8.11
weasyprint==59.0
scikit-learn==1.3.0
numpy==1.24.3
psycopg2-binary==2.9.7
gunicorn==21.2.0
email-validator==2.0.0
PyJWT==2.8.0
PyMuPDF==1.26.4  # For PDF annotation functionality

# Pour installation locale simple sans WeasyPrint (si problèmes):
# pip install flask flask-sqlalchemy flask-login flask-wtf
# pip install werkzeug requests pypdf2 python-docx
# pip install scikit-learn numpy python-dotenv
